chore: move nft did vector to __tests__
