+++ src/__tests__/nft-did-vector.ts
